Revert "Merge 'Compaction resharding tasks' from Aleksandra Martyniuk"
This reverts commit 2a58b4a39af74966a66d3563ea2d7b73e5b0c308, reversing
changes made to dd631690778e913029e95e14721d043bffa647b8.

After patch 87c8d63b7ac4a5e89c1bf2e21b970e3d2ffde593,
table_resharding_compaction_task_impl::run() performs the forbidden
action of copying a lw_shared_ptr (_owned_ranges_ptr) on a remote shard,
which is a data race that can cause a use-after-free, typically manifesting
as allocator corruption.

Note: before the bad patch, this was avoided by copying the _contents_ of the
lw_shared_ptr into a new, local lw_shared_ptr.

Fixes #14475
Fixes #14618
