+++ compaction/task_manager_module.hh
@@ -16,11 +16,6 @@
 namespace sstables {
 class sstable_directory;
 }
-
-namespace replica {
-class reshard_shard_descriptor;
-}
-
 namespace compaction {
 
 class compaction_task_impl : public tasks::task_manager::task::impl {
@@ -548,82 +543,6 @@ public:
         , _creator(std::move(creator))
         , _filter(std::move(filter))
         , _total_shard_size(total_shard_size)
-    {}
-
-    virtual tasks::is_internal is_internal() const noexcept override;
-protected:
-    virtual future<> run() override;
-};
-
-
-class resharding_compaction_task_impl : public compaction_task_impl {
-public:
-    resharding_compaction_task_impl(tasks::task_manager::module_ptr module,
-            tasks::task_id id,
-            unsigned sequence_number,
-            std::string keyspace,
-            std::string table,
-            std::string entity,
-            tasks::task_id parent_id) noexcept
-        : compaction_task_impl(module, id, sequence_number, std::move(keyspace), std::move(table), std::move(entity), parent_id)
-    {
-        // FIXME: add progress units
-    }
-
-    virtual std::string type() const override {
-        return "resharding compaction";
-    }
-protected:
-    virtual future<> run() override = 0;
-};
-
-class table_resharding_compaction_task_impl : public resharding_compaction_task_impl {
-private:
-    sharded<sstables::sstable_directory>& _dir;
-    sharded<replica::database>& _db;
-    sstables::compaction_sstable_creator_fn _creator;
-    compaction::owned_ranges_ptr _owned_ranges_ptr;
-public:
-    table_resharding_compaction_task_impl(tasks::task_manager::module_ptr module,
-            std::string keyspace,
-            std::string table,
-            sharded<sstables::sstable_directory>& dir,
-            sharded<replica::database>& db,
-            sstables::compaction_sstable_creator_fn creator,
-            compaction::owned_ranges_ptr owned_ranges_ptr) noexcept
-        : resharding_compaction_task_impl(module, tasks::task_id::create_random_id(), module->new_sequence_number(), std::move(keyspace), std::move(table), "", tasks::task_id::create_null_id())
-        , _dir(dir)
-        , _db(db)
-        , _creator(std::move(creator))
-        , _owned_ranges_ptr(std::move(owned_ranges_ptr))
-    {}
-protected:
-    virtual future<> run() override;
-};
-
-class shard_resharding_compaction_task_impl : public resharding_compaction_task_impl {
-private:
-    sharded<sstables::sstable_directory>& _dir;
-    replica::database& _db;
-    sstables::compaction_sstable_creator_fn _creator;
-    compaction::owned_ranges_ptr _owned_ranges_ptr;
-    std::vector<replica::reshard_shard_descriptor>& _destinations;
-public:
-    shard_resharding_compaction_task_impl(tasks::task_manager::module_ptr module,
-            std::string keyspace,
-            std::string table,
-            tasks::task_id parent_id,
-            sharded<sstables::sstable_directory>& dir,
-            replica::database& db,
-            sstables::compaction_sstable_creator_fn creator,
-            compaction::owned_ranges_ptr owned_ranges_ptr,
-            std::vector<replica::reshard_shard_descriptor>& destinations) noexcept
-        : resharding_compaction_task_impl(module, tasks::task_id::create_random_id(), 0, std::move(keyspace), std::move(table), "", parent_id)
-        , _dir(dir)
-        , _db(db)
-        , _creator(std::move(creator))
-        , _owned_ranges_ptr(std::move(owned_ranges_ptr))
-        , _destinations(destinations)
     {}
 
     virtual tasks::is_internal is_internal() const noexcept override;
